Overall Improvements (#5)
* Add support for OHLC and resampling

* Add concurrent I/O support

* Add new tests

* Improve usage of annotations

This commit adds multiple annotation improvements
which make it possible to not import libraries
such as pandas at runtime (unless required).

* Add tests and fix detected bugs

This commit adds some important tests over the public API
such as testing the get_tickers and get_quotes methods.

As a result of adding said tests, several bugs were found
which are also fixed as part of this commit.

* Optimize get_tickers

This commit adds a simple optimization to avoid
updating tickers if no new tickers are found